Close ExecuteFetchAs holes
